requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
isort==5.13.2
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
